backend-repo/requirements.txt
Flask==2.3.3 
python-dotenv==1.0.0
azure-storage-blob==12.19.0
Flask-Cors==4.0.0
